# expression: implement vectorized evaluation for `builtinAesEncryptIVSig`